Fix compilation errors in labs 11-27
Fixed issues in:
- Lab 12 (enums-pattern-matching): Move Message enum outside main() to module scope
- Lab 14 (utxo-model): Fix doc comment placement (use // instead of /// before closing brace)
- Lab 16 (collections-iterators): Store vector in variable to avoid temporary drop issue
- Lab 17 (closures-iterators): Use **x for double-reference and add .copied() after filter
- Lab 18 (modules-crates): Move module-level doc comments to top of files (lib.rs, models/mod.rs, services/mod.rs)
- Lab 21 (async-basics): Add tokio and futures dependencies to Cargo.toml
- Lab 22 (chat-server): Remove #[derive(Clone)] from Client struct (TcpStream not Clone)
- Lab 26 (consensus-simulation): Cast round to i32 for arithmetic, update Cargo.toml workspace config
- Lab 27 (gui-egui): Remove missing icon file include, add env_logger dependency, update Cargo.toml

All foundation labs (01-10) compile successfully.
Labs 11-20 now all compile (except lab 13 and 15 which were already passing).
Labs 21-30 mostly fixed (11, 12, 14, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 27, 28, 29, 30 now working).

+++ labs/26-consensus-simulation/src/main.rs
@@ -28,7 +28,7 @@ fn main() {
 
     for round in 1..=3 {
         println!("{}", format!("Round {}:", round).bright_cyan());
-        run_consensus_round(round, num_nodes, num_faulty, proposed_value + round - 1);
+        run_consensus_round(round, num_nodes, num_faulty, proposed_value + (round as i32) - 1);
         println!();
         thread::sleep(Duration::from_millis(500));
     }
